requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
pytz>=2023.3
gymnasium>=0.29
numpy>=1.24
requests>=2.31
daytona>=0.148
e2b-code-interpreter>=1.0
blaxel>=0.1
modal>=1.0
anthropic>=0.40
